refactor: fold csv2json into the main binary
